send email on register done,added utils and changed packaging for service
