+++ src/main/java/com/user/auth/utils/EmailUtils.java
@@ -1,0 +1,24 @@
+package com.user.auth.utils;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Component;
+
+@Component
+public class EmailUtils {
+
+    @Autowired private JavaMailSender emailSender;
+
+
+
+    public void sendInvitationEmail(String to, String subject, String text,String from) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setFrom(from);
+        message.setTo(to);
+        message.setSubject(subject);
+        message.setText(text);
+        emailSender.send(message);
+    }
+}
