+++ src/main/java/com/user/auth/utils/UserAuthUtils.java
@@ -1,0 +1,17 @@
+package com.user.auth.utils;
+
+import org.springframework.stereotype.Component;
+
+@Component
+public class UserAuthUtils {
+
+    public String generateKey(int n) {
+        String AlphaNumericString = "ABCDEFGHIJKLMNOPQRSTUVWXYZ" + "0123456789" + "abcdefghijklmnopqrstuvxyz";
+        StringBuilder sb = new StringBuilder(n);
+        for (int i = 0; i < n; i++) {
+            int index = (int) (AlphaNumericString.length() * Math.random());
+            sb.append(AlphaNumericString.charAt(index));
+        }
+        return sb.toString();
+    }
+}
