+++ src/main/resources/db/INSERT.SQL
@@ -1,0 +1,3 @@
+use user_auth;
+insert into role(role) values('ADMIN');
+insert into role(role) values('SUPER-ADMIN');
